Working recursive merge of NodeJS BOM files
Merging multiple cycloneDX BOMs currently requires an external CLI tool
intended for such utility functions. See comment here:
https://github.com/CycloneDX/cyclonedx-node-module/issues/90#issuecomment-729724784

This takes the approach of generating an XML-format BOM anywhere in a
repository tree a package.json file and a non-empty node_modules folder
are found together. The implicit assumption here is that a yarn/npm
install action will install *ALL* packages for all
components/sub-projects of monorepos and the like.

If that is NOT the case for your project, the resultant BOM will
necessarily not be representative of the complete superset of packages
contained in your top-level project.